Allow some renovate bumps in release branches
This enables automatic version bumps for dependencies in release
branches that don't follow the same minor version as the main branch.
Unfortunately, this requires extra manual configuration per branch and
dependency. Additionally, using base branches will change the branch
name pattern used by Renovate Bot.

Use prTitleStrict for the main branch so that pull requests against it
won't have a "(main)" suffix. Disable it for all other branches so they
receive the suffix. As a proof of concept, add a rule to update Go 1.23
in the 1.32 release branch.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>